Bootstrap verification changed
Now we just run 'knife node list' and check if the node name is in the list.
Before, we installed a getting-started recipe at bootstrap, and we verified if the getting-started.txt file was in the node (through ssh) to check whether the node had been bootstrapped.
The new method is much faster.