If EC2 instance is created but does not get accessible, NPM creates a new one
NodeBootstrapper tries to ssh into the node during 400 seconds.
The value 400 was retrieved for EC2 instaces from our scalability experiments.
If in 400 sec, the NodeBootstrapper cannot ssh the node, NPM tries to create a new node.
If the new node does not work, NPM gives up and throws the NodeNotCreatedException.
We have tested the new approach with the MultipleAWSNodesCreationTest, that creates 25 instances.