Verifying if node update was OK
Now the NodeUpgrader class in NPM tries to verify if the node upgrade (running chef-client) was OK.
The strategy is verifying if the new recipes are listed in the recipes line of knife node show command after the upgrade.
But one drawback: this must work just to the first application of each recipe.
In the next applications, the verification will return true, even if things were not OK.