NPMImpl does not use cloud provider to rertieve nodes anymore
Previously, the NodesRegistry was only a cache structure,
but now it holds the nodes managed by the NPM.
The idea is that NPM manages only know the nodes created by it,
otherwise it could try to use a VM that I created to other purposes.
This change leads to a more controlled environment.