Fixing idle pool tests
The tests ordered the creation of one VM in the pool followed by fillThePool.
Since both invocations are asynchornous, sometimes fillThePool was processed before the one vm creation.
Therefor, verifying poolSize >= N is the right verification in the test, and not poolSize = N,
since when fillThePool is processed before, the pool finishes the test with N+1 nodes.